Made loading config faster
